requirements.txt
# pip install --force-reinstall -r requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
numpy
matplotlib
